Reworking PypeIt, ScienceImage, and Reduce
